Replace [-f FIRST] [-n NUM] with FIRST+NUM syntax, default to join 225.1.2.3
This patch replaces the [-f FIRST] [-n NUM] command line options with a
more intuitive GROUP+NUM argument syntax.  This simplifies the code a
lot and folds the parsing and validation into the same path as the rest
of the code.

Also, if no command line arguments are given mcjoin now defaults to join
the group 225.1.2.3

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>