Improve stats, counters per group, increase MAX groups and one socket per group
While improving the stats for the receiver, the complete handling of
groups had to be refactored.  Introducing a `struct gr` to hold the
group, `struct sockaddr_in`, the per-group counter ... and a socket.

The latter was added as an experiment to see how well things scale, but
it was also discovered that while binding the socket to a given port,
all joined multicast groups are received on all sockets in the process!
Only way around it seems to be to use a different port per group, and
per socket.

The best thing would be to use `recvmsg()` and filter on the dest addr,
if only one could increase the MAX groups per socket ...

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>