Fix stack overrun bug in while parsing multiple groups
mcjoin can act on multiple groups, these groups are input as command
line arguments.  The mechanism for G+N is always executed, but when
there is no +N and the next group is longer than the previous one
the stack is overrun in inet_ntop().

This patch lets mcjoin operate on a copy of the group from the command
line and inet_ntop() operate on a big enough buffer at the end.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>