Handle case when mcjoin is started befoe any IP address/route is up
In case mcjoin is started with an explicit interface, -i ifname, we
probe for the interface on the (first) default route.  If there is
no default route we cannot find an interface and previously b0rked.

This patch adds support for falling back to the first reasonable
interface that can be found, instad of b0rking.  We could improve
this to make sure there is only one such possible interface, and
otherwise exit listing these alternatives to the user.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>